Use munit in IntersectSuite
